Implement order stuff

+++ schema/db/order.go
@@ -17,10 +17,10 @@ func (Order) Fields() []ent.Field {
 	return []ent.Field{
 		field.String("order_state"),
 		field.Int("quantity"),
-		field.Int("updated_at").
-		SchemaType(map[string]string{
-			dialect.Postgres: "timestampz",
-		}),
+		field.Int64("updated_at").
+			SchemaType(map[string]string{
+				dialect.Postgres: "timestampz",
+			}),
 	}
 }
